:zap: feature(Automated DQA):   -Checks for a beamset to be loaded     -Generates QA Plans for the current beamset using a centered delta4      phantom     -Sends the QA Plan to the Delta4 Dicom Destination     -Waits for the user to export this plan to iDMS     -Asks user to load the transfer plan     -Generates a dqa for this transfer plan on the centered delta 4     -Sends the QA plan to delta 4     -prompts user to send plan to iDMS     -Exports using the ValidationPlan.ScriptableQADicomExport method.     -Copies pertinent data to the user clipboard.
